# gh-131791: Improve speed of `textwrap.dedent` by replacing `re`