--- a/Lib/textwrap.py
+++ b/Lib/textwrap.py
@@ -433,7 +433,7 @@ def dedent(text):
 
     lines = text.split("\n")
 
-    margin = os.path.commonprefix(tuple(filter(lambda x: x.lstrip(), lines)))
+    margin = os.path.commonprefix([line for line in lines if line.lstrip()])
     margin_len = len(margin) - len(margin.lstrip())
 
     return "\n".join([line[margin_len:] if line.lstrip() else "\n" if line.endswith("\n") else "" for line in lines])