--- a/Lib/textwrap.py
+++ b/Lib/textwrap.py
@@ -435,7 +435,7 @@ def dedent(text):
     margin = os.path.commonprefix([line for line in lines if line.strip()])
     margin_len = len(margin) - len(margin.lstrip())
 
-    return "\n".join([line[margin_len:] if line.strip() else "\n" if line and line[-1] == "\n" else "" for line in lines])
+    return "\n".join([line[margin_len:] if line.strip() else "" for line in lines])
 
 
 def indent(text, prefix, predicate=None):