--- a/Misc/NEWS.d/next/Library/2025-03-27-04-35-17.gh-issue-131792.UtGg3O.rst
+++ b/Misc/NEWS.d/next/Library/2025-03-27-04-35-17.gh-issue-131792.UtGg3O.rst
@@ -1,0 +1,2 @@
+:func: `textwrap.dedent` is now 4x faster than before for large inputs. Function now has the command argument to remove all common prefixes as well with `only_whitespaces` instead of just whitespaces.
+Patch by Marius Juston.