--- a/Misc/NEWS.d/next/Library/2025-03-27-04-35-17.gh-issue-131792.UtGg3O.rst
+++ b/Misc/NEWS.d/next/Library/2025-03-27-04-35-17.gh-issue-131792.UtGg3O.rst
@@ -1,2 +1,1 @@
-Optimized :func: ``textwrap.dedent``. It is now 4x faster than before for large inputs. Function now has the command argument to remove all common prefixes as well with ``only_whitespace`` instead of just whitespaces.
-Patch by Marius Juston.
+Optimized :func:`textwrap.dedent`. It is now 2x faster than before for large inputs.