@Marius-Juston 2025-03-27
Yeah, most of this is still AI-generated (except for the final lines, where list comprehension was used instead of for loops). 

After reviewing all the comments, I was clearly overeager with the performance increase that I had initially noticed and did not perform the necessary due diligence that I really should have done in the first place. 

I was really using AI to see if it would be able to find interesting or new ways to optimize the function as a test ( the use of the `os.path.commonprefix` which is C optimized was such an interesting example ).

It might be better for me to just close this PR, do some more edits offline, verify all the tests, etc.. and then perform the PR?