@Marius-Juston 2025-03-27
@StanFromIreland @eendebakpt @ZeroIntensity Thank you for the comments, I have updated the code to be non AI generated and still give faster performance. Passes all the tests from what I can see. The code is also much simpler.

Not quite sure what this docs build failure is if one of you could elaborate on what I should do to fix it?