@Marius-Juston 2025-03-28
> I would like to add test cases with much longer test strings than those that are given and also cases that are multi-lines and with multiple blocks of indented texts. I also want to know when a file is considered to be "large". 

Based on what a previous PR did in issue https://github.com/python/cpython/issues/107369 I followed the same method and validated the "large" file with `Objects/unicodeobject.c` (this has 16k lines) that I indent at different levels, 

```python
with open("../Objects/unicodeobject.c") as f:
    raw_text = f.read()

tests = []
test_indents = ["abc  \t", '', "    ", "\t",  ' \t  abc', ' \t  abc  \t  ', ' ' * 1000]

for indent_v in test_indents:
    text = indent(raw_text, indent_v)
    tests.append(text)
```

the timings given are based just on the computation of calling the `dedent` function as such:

```python
tests = []
test_names = []
test_names = ["abc  \t", '', "    ", "\t",  ' \t  abc', ' \t  abc  \t  ', "' ' * 1000"]
test_indents = ["abc  \t", '', "    ", "\t",  ' \t  abc', ' \t  abc  \t  ', ' ' * 1000]

for indent_v in test_indents:
    text = indent(raw_text, indent_v)

    tests.append(text)

# Basic indented text with empty lines
text = """
    def hello():
        print("Hello, world!")


    """

tests.append(text)
test_names.append("Basic indented text with empty lines")

# Text with mixed indentation and blank lines
text = """
    This is a test.

    Another line.

"""

tests.append(text)
test_names.append("Text with mixed indentation and blank lines")

# No indentation (edge case)
text = """No indents here.
Just normal text.

With empty lines."""

tests.append(text)
test_names.append("No indentation (edge case)")

# Only blank lines (should preserve them)
text = """


"""

tests.append(text)
test_names.append("Only blank lines")

# Edge case: No common prefix to remove
text = """hello
    world
"""

tests.append(text)
test_names.append("Edge case: No common prefix to remove")

# Edge case: Single indented line
text = """    Only one indented line"""

tests.append(text)
test_names.append("Edge case: Single indented line")

# Edge case: Single indented line
text = """    """

tests.append(text)
test_names.append("Edge case: Single indented line only")

# Edge case: Single indented line
text = """"""

tests.append(text)
test_names.append("Edge case: Empty text")

for text, name in zip(tests, test_names):
    print(f"========= Case: {name.encode('ascii')} =========")

    a = dedent(text)

    for func in [dedent_faster_v5, dedent_faster]:
        single_test = func(text)

        print(func.__name__, a == single_test)

        it = timeit.Timer(lambda: func(text))
        result = it.repeat(number=1_000)
        result.sort()
        print(f"{func.__name__} Min: {result[0]:.4f}msec")
```

> Also, what happens when nothing should be dedented? are we loosing something when we do dedent(dedent(text))?

When nothing should be dedented the only the empty lines are normalized as per the documentation of `dedent`.

> On the "how" this was solved, I really think it's not a good idea to use `os.path.commonpath` as it's purpose is entirely different and we shouldn't really on this kind of magic. In addition ther docs say:
> 
> > Raise [ValueError](https://docs.python.org/3/library/exceptions.html#ValueError) if paths contain both absolute and relative pathnames, if paths are on different drives, or if paths is empty. Unlike [commonprefix()](https://docs.python.org/3/library/os.path.html#os.path.commonprefix), this returns a valid path.

I am not using `os.path.commonpath` but rather [`os.path.commonprefix`](https://docs.python.org/3.13/library/os.path.html#os.path.commonprefix), which does not raise any exceptions and follows the following documentation:

> Return the longest path prefix (taken character-by-character) that is a prefix of all paths in list. If list is empty, return the empty string ('').

So I believe that using this function is okay

> For benchmarks, please use pyperf instead of showing the minimum. The minimum doesn't help here as we're interested in average time complexity, not best time complexity. What's interesting when doing benchmarks is also the geometric mean and comparisons done on non-DEBUG builds (possibly with LTO+PGO enabled) [I don't know which kind of builds the tests were done on].

I was using as reference the Issue https://github.com/python/cpython/issues/107369 so I was performing something similar to them. As for the build my python environment is an Anaconda 3.12; however, I will make sure to setup the comparisons in a more proper environment for validation. Is there a quick link for how to set that up or how would you like it to be setup? Or should I just be doing

```shell
./configure --with-pydebug && make -j
./python speedtest.py
```